+++ src/controllers/MailingListTypesController.php
@@ -98,7 +98,6 @@ class MailingListTypesController extends Controller
         $mailingListType->subscribeVerificationRequired = (bool)$request->getBodyParam('subscribeVerificationRequired', $mailingListType->subscribeVerificationRequired);
         $mailingListType->subscribeVerificationEmailSubject = $request->getBodyParam('subscribeVerificationEmailSubject', $mailingListType->subscribeVerificationEmailSubject);
         $mailingListType->subscribeVerificationEmailTemplate = $request->getBodyParam('subscribeVerificationEmailTemplate', $mailingListType->subscribeVerificationEmailTemplate);
-        $mailingListType->subscribeVerificationSuccessTemplate = $request->getBodyParam('subscribeVerificationSuccessTemplate', $mailingListType->subscribeVerificationSuccessTemplate);
         $mailingListType->subscribeSuccessTemplate = $request->getBodyParam('subscribeSuccessTemplate', $mailingListType->subscribeSuccessTemplate);
         $mailingListType->unsubscribeFormAllowed = (bool)$request->getBodyParam('unsubscribeFormAllowed', $mailingListType->unsubscribeFormAllowed);
         $mailingListType->unsubscribeVerificationEmailSubject = $request->getBodyParam('unsubscribeVerificationEmailSubject', $mailingListType->unsubscribeVerificationEmailSubject);